+++ library/common/jni/jni_interface.cc
@@ -26,7 +26,7 @@ JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void* reserved) {
 
 // JniLibrary
 
-static void jvm_on_engine_running(void* context) {
+static void jvm_on_engine_running(envoy_engine_t engine_handle, void* context) {
   if (context == nullptr) {
     return;
   }
@@ -36,8 +36,8 @@ static void jvm_on_engine_running(void* context) {
   jobject j_context = static_cast<jobject>(context);
   jclass jcls_JvmonEngineRunningContext = env->GetObjectClass(j_context);
   jmethodID jmid_onEngineRunning = env->GetMethodID(
-      jcls_JvmonEngineRunningContext, "invokeOnEngineRunning", "()Ljava/lang/Object;");
-  env->CallObjectMethod(j_context, jmid_onEngineRunning);
+      jcls_JvmonEngineRunningContext, "invokeOnEngineRunning", "(J)Ljava/lang/Object;");
+  env->CallObjectMethod(j_context, jmid_onEngineRunning, static_cast<jlong>(engine_handle));
 
   env->DeleteLocalRef(jcls_JvmonEngineRunningContext);
   // TODO(goaway): This isn't re-used by other engine callbacks, so it's safe to delete here.