+++ test/swift/apps/experimental/ViewController.swift
@@ -26,7 +26,7 @@ final class ViewController: UITableViewController {
       .enableInterfaceBinding(true)
       // swiftlint:disable:next line_length
       .addNativeFilter(name: "envoy.filters.http.buffer", typedConfig: "{\"@type\":\"type.googleapis.com/envoy.extensions.filters.http.buffer.v3.Buffer\",\"max_request_bytes\":5242880}")
-      .setOnEngineRunning { NSLog("Envoy async internal setup completed") }
+      .setOnEngineRunning { _ in NSLog("Envoy async internal setup completed") }
       .addStringAccessor(name: "demo-accessor", accessor: { return "PlatformString" })
       .setEventTracker { NSLog("Envoy event emitted: \($0)") }
       .build()